Move stuck task detection into _periodic_database_sync
